refactor: Extract screenshot utils
